+++ src/BinanceBot.Market/Abstracts/BaseMarketBot.cs
