+++ src/BinanceBot.Market/Abstracts/IMarketBot.cs
