+++ src/BinanceBot.Market/Abstracts/IMarketStrategy.cs
