+++ src/BinanceBot.Market/Utility/QuoteExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Binance.Net.Enums;
+using BinanceBot.Market.Core;
 
 namespace BinanceBot.Market.Utility
 {
@@ -8,7 +9,8 @@ namespace BinanceBot.Market.Utility
     {
         public static IEnumerable<Quote> ToQuotes(this IDictionary<decimal, decimal> source, OrderSide direction)
         {
-            return source?.Select(s => new Quote(s.Key, s.Value, direction));
+            return source?
+                .Select(s => new Quote(s.Key, s.Value, direction));
         }
     }
 }
